Cron and Email refactor
Moved crons to their own folder, had to move some email stuff around to deal with the changes.
