+++ sciencelabs/sciencelabs_controller.py
@@ -1,37 +1,8 @@
 import random
 import string
 
-from flask import request, Response, abort, make_response, redirect
+from flask import abort
 from flask import session as flask_session
-from functools import wraps
-
-from sciencelabs import app
-
-
-def check_auth(username, password):
-    """This function is called to check if a username /
-    password combination is valid.
-    """
-    return username == app.config['LAB_LOGIN']['username'] and password == app.config['LAB_LOGIN']['password']
-
-
-def authenticate():
-    """Sends a 401 response that enables basic auth"""
-    return Response(
-        'Could not verify your access level for that URL.\n'
-        'You have to login with proper credentials', 401,
-        {'WWW-Authenticate': 'Basic realm="Login Required"'})
-
-
-def requires_auth(f):
-    @wraps(f)
-    def decorated(*args, **kwargs):
-        auth = request.authorization
-        if not auth or not check_auth(auth.username, auth.password):
-            return authenticate()
-        return f(*args, **kwargs)
-
-    return decorated
 
 
 class ScienceLabsController(object):
